+++ game/src/main/scala/io/aigar/score/ScoreModification.scala
@@ -2,5 +2,5 @@ package io.aigar.score
 
 case class ScoreModification(
   player_id: Int,
-  value: Int
+  value: Float
 )
